requirements_aws.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyJWT==2.8.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
boto3==1.34.0
botocore==1.34.0